feat: dividindo a autenticação em um controller e uma rota
