+++ main.py
@@ -1,5 +1,6 @@
 from fastapi import FastAPI
 from router.routes_usuario import router as usuario_router
+from router.routes_auth import router as routes_auth
 from config.helper import add_cors
 
 app = FastAPI()
@@ -7,6 +8,7 @@ app = FastAPI(debug=True)
 add_cors(app)
 
 app.include_router(usuario_router, tags=["usuarios"])
+app.include_router(routes_auth, tags=["login"])
 
 if __name__ == "__main__":
     import uvicorn
